+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.14.5] - 2025-09-26
+### Fixed
+- **Vial Search:** fixed bug where alchemist dedication would be prompted to add vials after 10+ minutes. 
+- **Powerful Alchemy:** fixed bug where it would send item description to chat before modifying the DC. 
+
 ## [2.14.4] - 2025-09-23
 ### Changed
 - **Formula Search:** updated the search on the formula page to also seach traits, and description. Credit to [crash1115](https://github.com/crash1115) for the suggestion!